+++ source/lambda/swap_secrets/index.js
@@ -11,8 +11,8 @@
  *  and limitations under the License.                                                                                *
  *********************************************************************************************************************/
  
- const aws = require('aws-sdk');
- const secretsmanager = process.env.METRICS == "true" ? new aws.SecretsManager({customUserAgent: process.env.SOLUTION_IDENTIFIER}) : new aws.SecretsManager();
+ const { SecretsManager } = require("@aws-sdk/client-secrets-manager");
+ const secretsmanager = process.env.METRICS == "true" ? new SecretsManager({customUserAgent: process.env.SOLUTION_IDENTIFIER}) : new SecretsManager();
  
  
  
@@ -29,7 +29,7 @@
          SecretId: temporaryKeyName
      };
  
-     let responseSecret = await secretsmanager.getSecretValue(params).promise();
+     let responseSecret = await secretsmanager.getSecretValue(params);
      console.log(responseSecret);
  
      const temporarySecretAsJson = JSON.parse(responseSecret.SecretString);
@@ -41,7 +41,7 @@
          SecretId: primaryKeyName
      };
  
-     responseSecret = await secretsmanager.getSecretValue(params).promise();
+     responseSecret = await secretsmanager.getSecretValue(params);
      
      const primarySecretAsJson = JSON.parse(responseSecret.SecretString);
      const primarySecretKeyName = Object.keys(primarySecretAsJson)[0];
@@ -55,7 +55,7 @@
          SecretId: secondaryKeyName, 
          SecretString: JSON.stringify(objectSecondary)
      };
-     await secretsmanager.putSecretValue(params).promise();
+     await secretsmanager.putSecretValue(params);
  
      const objectPrimary = {};
      objectPrimary[temporarySecretKeyName] = temporarySecretKeyValue;
@@ -66,7 +66,7 @@
          SecretString: JSON.stringify(objectPrimary)
      };
  
-     await secretsmanager.putSecretValue(params).promise();
+     await secretsmanager.putSecretValue(params);
  
      const objectTemporary = {};
      objectTemporary["INITIALIZED_KEY"] = "INITIALIZED_VALUE";
@@ -77,7 +77,7 @@
          SecretString: JSON.stringify(objectTemporary)
      };
 
-     await secretsmanager.putSecretValue(params).promise();
+     await secretsmanager.putSecretValue(params);
 
  
      return "OK";